+++ CHANGES.rst
@@ -1,13 +1,38 @@
 Changelog
 =========
 
+0.8.0rc1
+^^^^^^^^
+:release-date: 2017-08-07
+
+- Add support for sudo
+- Change default command name to `pywifi`, to evade conflicts with other packages
+- Add attribute `frequency_norm` which is the normalized frequency, aka frequency band
+  as one would find on packaging of WiFi hardware. Only 2.4Ghz and 5Ghz bands are
+  supported.
+- Fix some typo's in this document and create some more (probably)
+
 0.4.0
 ^^^^^
-:release-date: unknown
+:release-date: 2017-07-17
 
+- Internal release only
 - Make the wifi command name configurable (#55 - thanks yourealwaysbe)
 - Add a __main__.py so that wifi can be invoked using python -mwifi
 - Fix argument parsing so that scan is the default argument even with options passed
+- Add preliminary EAP support (#63 - thanks irgusite)
+- Update test suite and overall PEP8 formatting
+- Fix WPA detection
+- Fix relative signal level parsing to not rely on stable dictionary order
+- Centralize code to reformat a passphrase to a psk
+- Add ability to generate a `wpa_supplicant` network configuration from a cell
+- Add parsing for:
+  - Group cipher
+  - Pairwise ciphers
+  - Authentication suites
+  Please note that it's likely to be refactored into a separate object, as more then
+  one of these triplets may exist for a given cell.
+- Add logging
 
 0.3.8
 ^^^^^